+++ example/lib/counter/show_counter.dart
@@ -1,9 +1,10 @@
+import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:example/counter/counter.model.dart';
 import 'package:iconnect/iconnect.dart';
 
 class ShowCounterOrigin extends StatelessWidget {
-  ShowCounterOrigin({Key? key, this.value = 0}) : super(key: key);
+  const ShowCounterOrigin({Key? key, this.value = 0}) : super(key: key);
   final int value;
 
   @override
@@ -14,7 +15,9 @@ class ShowCounterOrigin extends StatelessWidget {
 
 Widget showCounter() {
   return Builder(builder: (context) {
-    print('ShowCounter build ');
+    if (kDebugMode) {
+      print('ShowCounter build ');
+    }
     //listen(context, first);
     listen<CounterModel>(context);
     return ShowCounterOrigin(value: first.value);
@@ -23,7 +26,9 @@ Widget showCounter() {
 
 Widget showCounterSecond() {
   return Builder(builder: (context) {
-    print('ShowCounterSecond build ');
+    if (kDebugMode) {
+      print('ShowCounterSecond build ');
+    }
     listen<CounterModel>(context, 'second');
     return ShowCounterOrigin(value: second.value);
   });
@@ -31,21 +36,24 @@ Widget showCounterSecond() {
 
 Widget showCounterAsync() {
   return Builder(builder: (context) {
-    print('ShowCounterSecond build ');
+    if (kDebugMode) {
+      print('ShowCounterSecond build ');
+    }
     listen(context, 'third');
     switch (third.snapshot.connectionState) {
       case ConnectionState.waiting:
-        return CircularProgressIndicator();
+        return const CircularProgressIndicator();
       default:
-        if (third.snapshot.hasError)
+        if (third.snapshot.hasError) {
           return Text(
             'Error: ${(third.snapshot.error as CounterError).message}',
-            style: TextStyle(color: Colors.red),
+            style: const TextStyle(color: Colors.red),
           );
-        else if (third.snapshot.hasData)
+        } else if (third.snapshot.hasData) {
           return ShowCounterOrigin(value: third.snapshot.data!.value);
-        else
+        } else {
           return ShowCounterOrigin(value: third.value);
+        }
     }
   });
 }