+++ example/lib/home/home.dart
@@ -1,7 +1,8 @@
 import 'package:flutter/material.dart';
 
 class MyHomePage extends StatelessWidget {
-  MyHomePage({Key? key, this.title, this.child, this.action}) : super(key: key);
+  const MyHomePage({Key? key, this.title, this.child, this.action})
+      : super(key: key);
 
   // This widget is the home page of your application. It is stateful, meaning
   // that it has a State object (defined below) that contains fields that affect
@@ -29,7 +30,7 @@ class MyHomePage extends StatelessWidget {
           // the App.build method, and use it to set our appbar title.
           title: Text(title!),
         ),
-        body: this.child,
-        floatingActionButton: this.action);
+        body: child,
+        floatingActionButton: action);
   }
 }