+++ lib/iconnect.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 
 import 'src/store.dart';
-export 'src/mixin.dart';
+//export 'src/mixin.dart';
 
 //two global function:provider and connect
 //one mixin for model:IConnect
@@ -10,6 +10,5 @@ get provider => Store.instance.provider;
 //if don't define model with mixin IConnect,could use these four function,for register,listen,dispatch
 Function<T>(BuildContext, [String?]) get listen => Store.instance.listen;
 void Function<T>([String? key]) get dispatch => Store.instance.dispatch;
-T Function<T>(T, {void Function<T>(T)? dispose, String? key}) get register =>
-    Store.instance.register;
+T Function<T>(T, [String? key]) get register => Store.instance.register;
 void Function<T>(String?) get unregister => Store.instance.unRegister;